move tests to unit/pipeline/test_task.py
